+++ src/lib.rs
@@ -1,3 +1,6 @@
 pub mod json;
+pub mod redirect;
+pub mod web;
 
 pub use json::Json;
+pub use redirect::Redirect;