+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased - 2021-xx-xx
 - Added reduced `Json` extractor with const generic size limit.
+- Added `Redirect` service.
 
 # 0.1.0
 - Empty crate.